Fix func tlbwrite
Sửa tham số kiểu int thành BYTE

+++ src/cpu-tlb.c
@@ -120,7 +120,7 @@ int tlbread(struct pcb_t * proc, uint32_t source,
  *@destination: index of destination register
  *@offset: destination address = [destination] + [offset]
  */
-int tlbwrite(struct pcb_t * proc, int data,
+int tlbwrite(struct pcb_t * proc, BYTE data,
              uint32_t destination, uint32_t offset)
 {
   int val;
